Fix mail error
This changes two things in the mail class.

1. If the html template was missing, it would tell you the txt template
was missing. It now tells you that the html template is missing.

2. If the type is set to only text or html, don't fail if the other
template is missing. Example, if set to only use text templates, don't
fail if the html template does not exist. If the type is set to both,
or if the specific type of template is missing, it will still complain.