// Small CSS fix for sale-label color
When using theme6.css the new-label and sale-label are sharing the same color (#719f41). It is not looking good and does not looks like the default theme which in there the colors of the labels are different from one and each other .
This small fix will add different color for each label.