[-] BO : fixed bug while use filter with enter key
If you use filter with enter key in customer or some admin pages.
And there is first img tag has onclick event like this getE('submitFiltercustomer').value=2 in form tag, that onclick event is fired. Because of that you will move 2nd page with blank result.
This is bug with some browser. It can be prevented with e.preventDefault() function.