CO: Calculate carrier price on the real order price
Calculate the carrier price with the REAL price customer is gonna pay, so have to use the BOTH_WITHOUT_SHIPPING const.

By now the Discount is not use in the calculation.
For exemple if you have a specific carrier price range for order > 50€ => free shipping. If the customer buy an 60€ item and use a -20% discount, customer gonna pay 48€ AND have the free shipping.

I think it's not the expected functionment