Improved Dispatcher: Get request URI from Symfony's Request (#9065)
* Dispatcher: Get request URI from Symfony's Request

Refactored method and mainly: get request URI from Symfony's request object. It's more reliable and future proof. I really like the idea to "get rid of the home-made framework which is not our core business features" from http://build.prestashop.com/news/prestashop-1-7-and-symfony/. Hope this is ok. Are there plans to rebuild the Dispatcher class with Symfony components by the way?

* Use Symfony's request to build request URI

Leave it optional to inject a Request object (to not break existing code) to the dispatcher. If not injected create one and save it in a property. This way the Dispatcher has access to the Symfony Request object and can build the request URI from it (and is able to use it somewhere else too or pass it around)

* doc

* spaces