Add module main class parse error to log message
During module installation the module main class is parsed to validated
its syntax. In case of a parse error an error message is logged. But the
log message does not include the actual parse error. The parse error is
entirely disregarded.

Adding the parse error to the log message helps to locate the exact
cause of why module installation failed.