Remove html escaping of errors
HTML escaping of errors were implemented in `PS 1.7.5.0-rc.1` [see commit](https://github.com/PrestaShop/PrestaShop/commit/8aa170e110b06c6ce6b8631eefa304f6206cfd7b). HTML escaping has not been implemented for any of the other message types `informations`, `confirmations` or `warnings`.

HTML escaping has the benefit of protecting against XSS injections but there is no protection in only adding this to one type of messages and this has a major drawback when plugins are trying to show HTML formatted error messages which a lot of plugins do. With the HTML escaping then it won't be able to implement lists, links or other helpful parts to a given error.

This PR removed the HTML escaping of errors to unify the behavior of all message types plus solving the issue that links cannot be included in error messages anymore.