Use scaled viewport for all product pages (incl. catalog)
Without a downscaled viewport OR removed gutters, the page will scroll horizontally.
I cannot remove gutters on the product page because it's not responsive. That page is also too crowded, so it needs to be downscaled.
I had to add a CSS exception for the product page so that gutters are not removed in that page.
Problem is, from CSS code I cannot distinguish between the product page and the catalog page... so gutters are NOT removed from the catalog page as well.
This forces me to apply downscaling to the catalog page as well in order to prevent it from scrolling horizontally.