Show free shipping indication only if appropriate
The problem is if you are not logged in or need postal codes to determine shipping, you always see shipping as "Free" in your cart, which really means "not calculated yet". It is bad for customer psychology to see "Free" and after they enter more information to get a price.

The result of this change is if there is shipping calculated then it will simply show it. If the system cannot determine shipping yet for any reason then you get "Not Calculated". Finally, if you really have free shipping then you get the indication "Free".